Add l.lib.testing.
