+++ landscape/lib/config.py
@@ -1,0 +1,2 @@
+
+from landscape.deployment import BaseConfiguration  # NOQA
